Use TSX for share dialog, split it into a separate file
